Switched post_gen hook from os to pathlib

+++ maestro-service/hooks/post_gen_project.py
@@ -14,10 +14,11 @@ print(REMOVE_PATHS)
 
 # Loop through paths and remove them if they exist
 for path_str in REMOVE_PATHS:
-    path = Path(path_str.strip())
-    print(path)
-    if path.exists():
-        if path.is_file():
-            path.unlink()  # Remove file
-        elif path.is_dir():
-            path.rmdir()  # Remove directory
+    if path_str:
+        path = Path(path_str.strip())
+        print(path)
+        if path.exists():
+            if path.is_file():
+                path.unlink()  # Remove file
+            elif path.is_dir():
+                path.rmdir()  # Remove directory
